refactor: share confirm page
